Release v2.2.0 (Windows/Linux interaction parity + capture/drag visual feedback)
Squash of branch v2: full-code-audit hardening sprint, ADR-052..060 —
Main-toolbar exclusion + toolbar table repair, MIDI toolbars first-class,
Win32/Linux capture-by-click + drag-to-dock revived and verified live on
VMs, 4-strip overlay frame (hover outline + drop-zone feedback) with
hybrid title-bar aiming and zone debounce, collapsible nav bar,
single-tab tab-bar collapse, silent-failure surfacing, Linux parity
(links/chrome/dark-auto/clamp), updater unload hardening, unified
parsers + tests (329 assertions), -Werror CI gate made real.

+++ cpp/src/split_tree.cpp
@@ -234,7 +234,7 @@ void SplitTree::SplitRectV(const RECT& parent, float ratio,
                             RECT& outLeft, RECT& outSplitter, RECT& outRight)
 {
   int w = parent.right - parent.left;
-  int splitX = parent.left + (int)(w * ratio);
+  int splitX = parent.left + (int)((float)w * ratio);
   splitX = clamp_i(splitX, parent.left + MIN_PANE_SIZE, parent.right - MIN_PANE_SIZE - SPLITTER_WIDTH);
 
   outLeft = parent;
@@ -252,7 +252,7 @@ void SplitTree::SplitRectH(const RECT& parent, float ratio,
                             RECT& outTop, RECT& outSplitter, RECT& outBottom)
 {
   int h = parent.bottom - parent.top;
-  int splitY = parent.top + (int)(h * ratio);
+  int splitY = parent.top + (int)((float)h * ratio);
   splitY = clamp_i(splitY, parent.top + MIN_PANE_SIZE, parent.bottom - MIN_PANE_SIZE - SPLITTER_WIDTH);
 
   outTop = parent;
